mcp: host /mcp alongside other http routes
