+++ shiksha-api/app-service/app/config.py
@@ -10,7 +10,6 @@ class Settings(BaseSettings):
     debug: bool = False
     host: str = "0.0.0.0"
     port: int = 8000
-    transport: str = "http"
 
     # Logging Configuration
     log_level: str = "INFO"
