+++ shiksha-api/app-service/app/main.py
@@ -1,9 +1,104 @@
+from fastapi import FastAPI, HTTPException
+from fastapi.middleware.cors import CORSMiddleware
+from contextlib import asynccontextmanager
+
+from fastmcp import FastMCP
+from fastmcp.server.http import StarletteWithLifespan
 from app.config import settings
+from app.routers import chat_router, chat_router_mcp, question_paper_router
+
+
+@asynccontextmanager
+async def lifespan(app: FastAPI):
+    """Lifespan context manager for proper startup and shutdown"""
+    # Startup
+    assert isinstance(app.state.MCP_APP, StarletteWithLifespan)
+    async with app.state.MCP_APP.lifespan(app):
+        yield
+    # Shutdown
+    from app.services.general_chat_service import GENERAL_CHAT_SERVICE_INSTANCE
+    from app.services.lesson_chat_service import LESSON_CHAT_SERVICE_INSTANCE
+    from app.services.question_paper_service import QUESTION_PAPER_SERVICE_INSTANCE
+
+    try:
+        await GENERAL_CHAT_SERVICE_INSTANCE.cleanup()
+        await LESSON_CHAT_SERVICE_INSTANCE.cleanup()
+        await QUESTION_PAPER_SERVICE_INSTANCE.cleanup()
+    except Exception as e:
+        print(f"Error during cleanup: {e}")
+
+
+app = FastAPI(
+    title=settings.app_name,
+    version=settings.version,
+    description="AI-powered educational chat API for Shiksha platform",
+    debug=settings.debug,
+    lifespan=lifespan,
+    redirect_slashes=False,
+)
+
+mcp = FastMCP(
+    name=settings.app_name,
+    instructions="AI-powered educational chat API for Shiksha platform",
+    host=settings.host,
+    port=settings.port,
+    debug=settings.debug,
+    stateless_http=True
+)
+app.state.MCP_APP = mcp.http_app(path="/mcp")
+app.mount("/", app.state.MCP_APP)
+
+# CORS middleware
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["*"],  # Configure appropriately for production
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+# Include routers
+app.include_router(chat_router)
+app.include_router(question_paper_router)
+
+# Include MCP tools
+chat_router_mcp(mcp)
+
+@mcp.tool(name="version", description="Get the version")
+async def mcp_version() -> str: return settings.version
+
+@mcp.tool(name="app_name", description="Get the app name")
+async def mcp_app_name() -> str: return settings.app_name
+
+@mcp.tool(name="health", description="Get the health status")
+async def mcp_health() -> str: return "healthy"
+
+
+@app.get("/")
+async def root():
+    """Root endpoint for health check"""
+    return {
+        "message": "Welcome to Shiksha Copilot API",
+        "version": settings.version,
+        "status": "healthy",
+    }
+
+
+@app.get("/health")
+async def health_check():
+    """Health check endpoint"""
+    return {"status": "healthy", "service": settings.app_name}
+
+
+# Global exception handler
+@app.exception_handler(Exception)
+async def global_exception_handler(request, exc):
+    return HTTPException(status_code=500, detail="Internal server error")
+
 
 if __name__ == "__main__":
-    if settings.transport == "http":
-        import app.serve_http
-    elif settings.transport in ("mcp-http", "mcp-sse", "mcp-stdio"):
-        import app.serve_mcp
-    else:
-        raise ValueError("Unexpected transport: " + settings.transport)
+    import uvicorn
+
+    uvicorn.run(
+        "app.main:app", host=settings.host, port=settings.port, reload=settings.debug
+    )
