+++ shiksha-api/app-service/pyproject.toml
@@ -13,7 +13,7 @@ uvicorn = { extras = ["standard"], version = "^0.35.0" }
 pydantic = "^2.5.0"
 python-multipart = "^0.0.9"
 python-dotenv = "^1.0.0"
-httpx = "^0.27.1"
+httpx = "^0.28.1"
 aiofiles = "^23.2.1"
 pyyaml = "^6.0.1"
 autogen-agentchat = "*"
@@ -22,7 +22,7 @@ aiohttp = "^3.9.0"
 vellm-rag-wrapper = { path = "app/rag-wrapper", develop = true }
 autogen-ext = { extras = ["azure", "openai"], version = "^0.6.4" }
 pydantic-settings = "^2.10.1"
-mcp = {extras = ["cli"], version = "^1.14.0"}
+fastmcp = "^2.13.0.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.4.2"
@@ -36,7 +36,6 @@ pre-commit = "^3.5.0"
 
 [tool.poetry.group.test.dependencies]
 pytest-mock = "^3.12.0"
-httpx = "^0.27.1"
 
 [build-system]
 requires = ["poetry-core"]
